logs and login are separated, no index
